Cached skip() generator values
the skip() generator was being called repeatedly in a loop where it was being re-evaluated per iteration, but does not have any dependencies on the values within the loop. Storing the results of the generator before using it in a loop saves hundreds of further unnecessary generator calls. This can be a potential bottleneck if several videos were to be downloaded.